requirements.txt
streamlit
tensorflow
tensorflow-hub
PyPDF2
python-docx
